[PB-2409] refactor rename team
